Move AggregatorTestCase to test framework
Allows plugins to use the test class.  Only potentially controversial point is that
`build()` had to be made public instead of protected... but the sole user of that API
(AbstractAggregatorBuilder) already exposed it as public so I think this is ok.
