+++ core/src/test/java/org/elasticsearch/search/aggregations/metrics/sum/SumAggregatorTests.java
@@ -37,7 +37,7 @@ import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.common.CheckedConsumer;
 import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.index.mapper.NumberFieldMapper;
-import org.elasticsearch.search.aggregations.AggregatorTestCase;
+import org.elasticsearch.test.AggregatorTestCase;
 
 import java.io.IOException;
 import java.util.Arrays;
